removed blueprints/auth.py
all user-related CRUD operations are now found in blueprints/users.py

NOTE: no data validation included
