+++ db/app.py
@@ -2,7 +2,6 @@ from flask import Flask
 from models import db
 from config import Config
 
-from blueprints.auth import auth_bp
 from blueprints.reports import reports_bp
 from blueprints.alerts import alerts_bp
 from blueprints.users import users_bp
@@ -14,7 +13,6 @@ def create_app():
     db.init_app(app)
 
     # Register blueprints
-    app.register_blueprint(auth_bp, url_prefix='/auth')
     app.register_blueprint(reports_bp, url_prefix='/reports')
     app.register_blueprint(alerts_bp, url_prefix='/alerts')
     app.register_blueprint(users_bp, url_prefix='/user')
